move DataSetException into correct package, remove deprecated package
